requirements.txt
streamlit
plotly.express
pandas
python-dotenv